+++ src/controllers/query.controller.ts
@@ -1,8 +1,7 @@
-import { randomUUID } from "node:crypto";
 import type { NextFunction, Request, Response } from "express";
 import type { QueryService } from "@/services";
 import { MessageRole } from "@/types/memory";
-import { loggers } from "@/utils/logger";
+import { streamEventsToSSE } from "@/utils/sse-stream";
 
 export class QueryController {
 	private queryService: QueryService;
@@ -64,72 +63,29 @@ export class QueryController {
 		} = req.body;
 		const userId = res.locals.userId;
 
-		res.writeHead(200, {
-			"Content-Type": "text/event-stream",
-			"Cache-Control": "no-cache",
-			Connection: "keep-alive",
-			"X-Accel-Buffering": "no", // nginx 버퍼링 비활성화
-		});
-		res.flushHeaders();
-		res.write(":ok\n\n");
-
-		const keepaliveInterval = setInterval(() => {
-			res.write(":keepalive\n\n");
-		}, 10000); // 10초마다 keepalive 전송
-
-		// 클라이언트 연결 끊김 감지
-		let aborted = false;
-		req.on("close", () => {
-			aborted = true;
-			loggers.intentStream.info("Client connection closed", {
-				threadId: currentThreadId,
-				userId,
-			});
-		});
-
-		let currentThreadId = threadId;
-		const stream = this.queryService.handleQuery(
-			{ type, userId, threadId, workflowId, title },
-			{ query, displayQuery },
-		);
-
-		try {
-			for await (const event of stream) {
-				if (aborted) {
-					break;
-				}
-
-				if (event.event === "thread_id") {
-					currentThreadId = event.data.threadId;
-				} else if (event.event === "thinking_process") {
-					// a2a 호출에 대해서는 데이터베이스에 추가하지 않기 위해 여기서 thread message에 기록
-					const thinkData =
-						await this.queryService.filterThinkingDataForStorage(event.data);
-					await this.queryService.addToThreadMessages(userId, currentThreadId, [
-						{
-							messageId: randomUUID(),
-							role: MessageRole.MODEL,
-							timestamp: Date.now(),
-							content: { type: "text", parts: [thinkData.title] },
-							metadata: {
-								isThinking: true,
-								thinkData,
-							},
-						},
-					]);
-				}
-
-				res.write(
-					`event: ${event.event}\ndata: ${JSON.stringify(event.data)}\n\n`,
+		await streamEventsToSSE(req, res, {
+			logLabel: "query stream",
+			userId,
+			setup: async () =>
+				this.queryService.handleQuery(
+					{ type, userId, threadId, workflowId, title },
+					{ query, displayQuery },
+				),
+			onThinkingProcess: async (currentThreadId, data) => {
+				// a2a 호출에 대해서는 데이터베이스에 추가하지 않기 위해 여기서 thread message에 기록
+				const thinkData =
+					await this.queryService.filterThinkingDataForStorage(data);
+				await this.queryService.addTextMessage(
+					userId,
+					currentThreadId,
+					MessageRole.MODEL,
+					thinkData.title,
+					{
+						isThinking: true,
+						thinkData,
+					},
 				);
-			}
-		} catch (error: unknown) {
-			const errMsg =
-				(error as Error)?.message || "Failed to handle query stream";
-			res.write(`event: error\ndata: ${errMsg}\n\n`);
-		} finally {
-			clearInterval(keepaliveInterval);
-			res.end();
-		}
+			},
+		});
 	};
 }
